move schema to models folder like other qiskit components
