+++ qiskit/compiler/run_config.py
@@ -7,23 +7,8 @@
 
 """Models for RunConfig and its related components."""
 
-from marshmallow.validate import Range
-
-from qiskit.validation import BaseModel, BaseSchema, bind_schema
-from qiskit.validation.fields import Boolean, Integer
-
-
-class RunConfigSchema(BaseSchema):
-    """Schema for RunConfig."""
-
-    # Required properties.
-    # None
-
-    # Optional properties.
-    shots = Integer(validate=Range(min=1))
-    max_credits = Integer(validate=Range(min=3, max=10))  # TODO: can we check the range
-    seed = Integer()
-    memory = Boolean()  # set default to be False
+from qiskit.compiler.models import RunConfigSchema
+from qiskit.validation import BaseModel, bind_schema
 
 
 @bind_schema(RunConfigSchema)
